[#29] refact : App.tsx -> Layout.tsx (teamsImage)
